// src/api/generated/model/signalDetectedEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { SignalDetectedEventType } from './signalDetectedEventType';
import type { SignalDetectedPayload } from './signalDetectedPayload';

export interface SignalDetectedEvent {
  id: string;
  type: SignalDetectedEventType;
  payload: SignalDetectedPayload;
  created_at: string;
  processed_at: string | null;
}